refactor: split House tests into separate test classes, improve code, add tests for constructor call with invalid value
